Another test of git commit tracking
Ignore this message, I made some unnecessary changes to try and see if git is tracking them or not. I'm having the hardest time getting this to cooperate, and it's confusing because it doesn't appear that I've changed anything in the way I am operating with git and GitHub. I'll be sure to create some sort of resource to fully document my proper git process once I've discovered it! Until then, onto making more mistakes!