Add exclamation point to title
This is simply another git test, I've discovered that git is working just fine in other repos, but not in the css-exercises repo. I suspect that it has something to do with creating a branch on css-exercises. This is a test to see if the problem is exclusive to the flex section.